+++ src/pytest_bdd/parser.py
@@ -9,14 +9,14 @@ from typing import Any, Dict, Iterable, List, Mapping, Optional, Sequence
 
 from .exceptions import StepError
 from .gherkin_parser import Background as GherkinBackground
-from .gherkin_parser import DataTable, ExamplesTable
+from .gherkin_parser import DataTable
 from .gherkin_parser import Feature as GherkinFeature
 from .gherkin_parser import GherkinDocument
 from .gherkin_parser import Scenario as GherkinScenario
 from .gherkin_parser import Step as GherkinStep
 from .gherkin_parser import Tag as GherkinTag
 from .gherkin_parser import get_gherkin_document
-from .types import GIVEN, STEP_TYPES, THEN, WHEN
+from .types import STEP_TYPES
 
 STEP_PARAM_RE = re.compile(r"<(.+?)>")
 COMMENT_RE = re.compile(r"(^|(?<=\s))#")
@@ -172,6 +172,7 @@ class ScenarioTemplate:
                 line_number=step.line_number,
                 keyword=step.keyword,
                 datatable=step.datatable,
+                docstring=step.docstring,
             )
             for step in self._steps
         ]
@@ -227,13 +228,21 @@ class Step:
     line_number: int
     indent: int
     keyword: str
+    docstring: str | None = None
     datatable: DataTable | None = None
     failed: bool = field(init=False, default=False)
     scenario: ScenarioTemplate | None = field(init=False, default=None)
     background: Background | None = field(init=False, default=None)
 
     def __init__(
-        self, name: str, type: str, indent: int, line_number: int, keyword: str, datatable: DataTable | None = None
+        self,
+        name: str,
+        type: str,
+        indent: int,
+        line_number: int,
+        keyword: str,
+        datatable: DataTable | None = None,
+        docstring: str | None = None,
     ) -> None:
         """Initialize a step.
 
@@ -250,6 +259,7 @@ class Step:
         self.line_number = line_number
         self.keyword = keyword
         self.datatable = datatable
+        self.docstring = docstring
 
     def __str__(self) -> str:
         """Return a string representation of the step.
@@ -346,12 +356,6 @@ class FeatureParser:
             List[Step]: A list of Step objects.
         """
 
-        def get_step_content(_gherkin_step: GherkinStep) -> str:
-            step_name = strip_comments(_gherkin_step.text)
-            if _gherkin_step.docstring:
-                step_name = f"{step_name}\n{_gherkin_step.docstring.content}"
-            return step_name
-
         if not steps_data:
             return []
 
@@ -360,25 +364,25 @@ class FeatureParser:
             raise StepError(
                 message=f"First step in a scenario or background must start with 'Given', 'When' or 'Then', but got {first_step.keyword}.",
                 line=first_step.location.line,
-                line_content=get_step_content(first_step),
+                line_content=first_step.text,
                 filename=self.abs_filename,
             )
 
         steps = []
         current_type = first_step.keyword.lower()
         for step in steps_data:
-            name = get_step_content(step)
             keyword = step.keyword.lower()
             if keyword in STEP_TYPES:
                 current_type = keyword
             steps.append(
                 Step(
-                    name=name,
+                    name=strip_comments(step.text),
                     type=current_type,
                     indent=step.location.column - 1,
                     line_number=step.location.line,
                     keyword=step.keyword.title(),
                     datatable=step.datatable,
+                    docstring=step.docstring.content if step.docstring else None,
                 )
             )
         return steps