+++ CHANGES.rst
@@ -6,7 +6,8 @@ Unreleased
 - Update documentation to clarify that `--gherkin-terminal-reporter` needs to be used with `-v` or `-vv`.
 - Drop compatibility with pytest < 7.0.0.
 - Continuation of steps using asterisks instead of And/But supported.
-- Added `datatable` argument for steps that contain a datatable.
+- Added `datatable` argument for steps that contain a datatable (see Data Tables section at https://cucumber.io/docs/gherkin/reference/)
+- Added `docstring` argument for steps that contain a docstring (see Doc Strings section at https://cucumber.io/docs/gherkin/reference/).
 
 8.0.0b1
 ----------